+++ domains/certificates/Libraries/CertificateEvents/CertificateCreated.cs
@@ -1,5 +1,5 @@
 using System;
-using System.Numerics;
+using CertificateEvents.Primitives;
 using EnergyOriginEventStore.EventStore.Serialization;
 
 namespace CertificateEvents;
@@ -8,24 +8,9 @@ namespace CertificateEvents;
 public record CertificateCreated(
     Guid CertificateId, // Guid.newGuid()
     string GridArea, // stamdata
-    Period Period, // EnergyMeasured.DateFrom
+    Period Period,
     Technology Technology, // stamdata
     byte[] OwnerPublicKey, // stamdata - meterpoint owners key
     ShieldedValue<string> ShieldedGSRN, //Group.commit(EnergyMeasured.GSRN)
     ShieldedValue<long> ShieldedQuantity //Group.commit(EnergyMeasured.Quantity)
 ) : EventModel;
-
-public record Period(
-    long DateFrom, // EnergyMeasured.DateFrom
-    long DateTo  // EnergyMeasured.DateTo
-);
-
-public record Technology(
-    string FuelCode, // stamdata
-    string TechCode  // stamdata
-);
-
-public record ShieldedValue<T>(
-    T Value, // stamdata
-    BigInteger R  // group.RandomR()
-);
