+++ domains/certificates/Libraries/CertificateEvents/Primitives/Period.cs
@@ -1,0 +1,6 @@
+namespace CertificateEvents.Primitives;
+
+public record Period(
+    long DateFrom, // EnergyMeasured.DateFrom
+    long DateTo  // EnergyMeasured.DateTo
+);
