+++ domains/certificates/Libraries/CertificateEvents/Primitives/ShieldedValue.cs
@@ -1,0 +1,8 @@
+using System.Numerics;
+
+namespace CertificateEvents.Primitives;
+
+public record ShieldedValue<T>(
+    T Value, // stamdata
+    BigInteger R  // group.RandomR()
+);
