+++ domains/certificates/Libraries/CertificateEvents/Primitives/Technology.cs
@@ -1,0 +1,6 @@
+namespace CertificateEvents.Primitives;
+
+public record Technology(
+    string FuelCode, // stamdata
+    string TechCode  // stamdata
+);
